[#42] common: Replace InvokeID function to vote.go source file
