+++ common/invoke.go
@@ -1,18 +1,8 @@
 package common
 
 import (
-	"github.com/nspcc-dev/neo-go/pkg/interop/crypto"
 	"github.com/nspcc-dev/neo-go/pkg/interop/runtime"
 )
-
-func InvokeID(args []interface{}, prefix []byte) []byte {
-	for i := range args {
-		arg := args[i].([]byte)
-		prefix = append(prefix, arg...)
-	}
-
-	return crypto.SHA256(prefix)
-}
 
 type IRNode struct {
 	PublicKey []byte
